List view now works, formatting is basic but works
Hello all, sorry it's been a while, I've been busy digging up the garden. Anyway, today I bring the app a step closer to being functional, we have a working menu bar, we have a working list of stations, all that's left to do is get the stations to play music, this should hopefully be done soon!